+++ Snabble/UI/Payment/PaymentMethod+UI.swift
@@ -16,12 +16,26 @@ extension PaymentMethod {
         }
     }
 
-    func processor(_ process: CheckoutProcess, _ cart: ShoppingCart, _ delegate: PaymentDelegate) -> UIViewController {
+    var processRequired: Bool {
         switch self {
-        case .cash: return CashCheckoutViewController(process, cart, delegate)
-        case .qrCode: return QRCheckoutViewController(process, cart, delegate)
-        case .encodedCodes: return EmbeddedCodesCheckoutViewController(process, cart, delegate)
+        case .cash, .qrCode: return true
+        case .encodedCodes: return false
         }
+    }
+
+    func processor(_ process: CheckoutProcess?, _ cart: ShoppingCart, _ delegate: PaymentDelegate) -> UIViewController? {
+        if self.processRequired && process == nil {
+            return nil
+        }
+
+        let processor: UIViewController
+        switch self {
+        case .cash: processor = CashCheckoutViewController(process!, cart, delegate)
+        case .qrCode: processor = QRCheckoutViewController(process!, cart, delegate)
+        case .encodedCodes: processor = EmbeddedCodesCheckoutViewController(process, cart, delegate)
+        }
+        processor.hidesBottomBarWhenPushed = true
+        return processor
     }
 }
 
@@ -49,7 +63,8 @@ public class PaymentProcess {
     /// if the checkout allows multiple payment methods, offer a selection
     /// otherwise, directly create the corresponding view controller for the selected payment mehthod
     ///
-    /// - Parameter completion: a closure called when the payment method has been determined. This is passed a `UIViewController` instance that the caller can present
+    /// - Parameter completion: a closure called when the payment method has been determined.
+    ///             This is passed a `UIViewController` instance that the caller can present
     public func start(completion: @escaping (UIViewController) -> () ) {
         let info = self.signedCheckoutInfo
         if info.checkoutInfo.paymentMethods.count > 1 {
@@ -65,27 +80,31 @@ public class PaymentProcess {
 
     func start(_ method: PaymentMethod, completion: @escaping (UIViewController) -> () ) {
         UIApplication.shared.beginIgnoringInteractionEvents()
-        self.showHud()
+        self.startBlurOverlayTimer()
 
-        self.signedCheckoutInfo.createCheckoutProcess(SnabbleUI.project, paymentMethod: method) { process, error in
+        self.signedCheckoutInfo.createCheckoutProcess(SnabbleUI.project, paymentMethod: method, timeout: 2) { process, error in
             self.hudTimer?.invalidate()
             self.hudTimer = nil
             UIApplication.shared.endIgnoringInteractionEvents()
             self.hideBlurOverlay()
-            if let process = process {
-                let processor = method.processor(process, self.cart, self.delegate)
-                processor.hidesBottomBarWhenPushed = true
+            if let process = process, let processor = method.processor(process, self.cart, self.delegate) {
                 completion(processor)
             } else {
                 let handled = self.delegate.handlePaymentError(error)
                 if !handled {
-                    self.delegate.showInfoMessage("Snabble.Payment.errorStarting".localized())
+                    if method == .encodedCodes, let processor = method.processor(nil, self.cart, self.delegate) {
+                        // continue anyway
+                        completion(processor)
+                        self.retryCreatingMissingCheckout()
+                    } else {
+                        self.delegate.showInfoMessage("Snabble.Payment.errorStarting".localized())
+                    }
                 }
             }
         }
     }
 
-    private func showHud() {
+    private func startBlurOverlayTimer() {
         self.hudTimer = Timer.scheduledTimer(withTimeInterval: 0.25, repeats: false) { _ in
             self.showBlurOverlay()
         }
@@ -93,6 +112,15 @@ public class PaymentProcess {
 
     func track(_ event: AnalyticsEvent) {
         self.delegate.track(event)
+    }
+
+    // retry creating the checkout info / checkout process that is potentially missing
+    private func retryCreatingMissingCheckout() {
+        DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
+            self.cart.createCheckoutInfo(SnabbleUI.project) { info, _ in
+                info?.createCheckoutProcess(SnabbleUI.project, paymentMethod: .encodedCodes) { _,_ in }
+            }
+        }
     }
 
     private var blurView: UIView?