+++ Snabble/UI/Scanner/ScannerViewController.swift
@@ -11,6 +11,20 @@ public protocol ScannerDelegate: AnalyticsDelegate, MessageDelegate {
     func closeScanningView()
 }
 
+extension ScanFormat {
+    var avType: AVMetadataObject.ObjectType {
+        switch self {
+        case .ean8: return .ean8
+        case .ean13: return .ean13
+        case .code128: return .code128
+        case .itf14: return .itf14
+        case .code39: return .code39
+        case .qr: return .qr
+        case .dataMatrix: return .dataMatrix
+        }
+    }
+}
+
 public class ScannerViewController: UIViewController {
 
     @IBOutlet private weak var spinner: UIActivityIndicatorView!
@@ -31,21 +45,19 @@ public class ScannerViewController: UIViewController {
     
     private var hiddenConfirmationOffset: CGFloat = 310
     private var keyboardObserver: KeyboardObserver!
-    private var objectTypes: [AVMetadataObject.ObjectType] = [ .ean8, .ean13, .code128 ]
+    private var objectTypes = [AVMetadataObject.ObjectType]()
     private weak var delegate: ScannerDelegate!
     private var timer: Timer?
 
-    public init(_ productProvider: ProductProvider, _ cart: ShoppingCart, _ shop: Shop, delegate: ScannerDelegate, objectTypes: [AVMetadataObject.ObjectType]? = nil) {
-        self.productProvider = productProvider
+    public init(_ project: Project, _ cart: ShoppingCart, _ shop: Shop, delegate: ScannerDelegate) {
+        self.productProvider = SnabbleAPI.productProvider(for: project)
         self.shoppingCart = cart
         self.shop = shop
+        self.objectTypes = project.scanFormats.map { $0.avType }
 
         super.init(nibName: nil, bundle: Snabble.bundle)
 
         self.delegate = delegate
-        if let objectTypes = objectTypes {
-            self.objectTypes = objectTypes
-        }
 
         self.title = "Snabble.Scanner.title".localized()
         self.tabBarItem.image = UIImage.fromBundle("icon-scan")
@@ -123,10 +135,6 @@ public class ScannerViewController: UIViewController {
         self.keyboardObserver = KeyboardObserver(handler: self)
     }
 
-    override public func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-    }
-
     override public func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
 
@@ -145,10 +153,12 @@ public class ScannerViewController: UIViewController {
         self.infoView.isHidden = true
     }
 
-    /// reset `productProvider` and `shoppingCart` when switching between projects
-    public func reset(_ productProvider: ProductProvider, _ cart: ShoppingCart) {
-        self.productProvider = productProvider
+    /// reset `project` and `shoppingCart` when switching between projects
+    public func reset(_ project: Project, _ cart: ShoppingCart) {
+        self.productProvider = SnabbleAPI.productProvider(for: project)
         self.shoppingCart = cart
+        self.objectTypes = project.scanFormats.map { $0.avType }
+        self.scanningView?.setObjectTypes(self.objectTypes)
 
         // avoid camera permission query if this is called before we've ever been on-screen
         if self.scanningView != nil {
@@ -156,7 +166,7 @@ public class ScannerViewController: UIViewController {
             self.navigationController?.popToRootViewController(animated: false)
         }
     }
-    
+
     // MARK: - scan confirmation views
     
     private func showConfirmation(for product: Product, _ code: String) {
@@ -404,8 +414,8 @@ extension ScannerViewController {
                     return
                 }
             }
-            
-            self.productProvider.productByWeighItemId(ean.codeForLookup) { product, error in
+
+            self.productProvider.productByWeighItemId(ean.codeForLookup, self.shop.id) { product, error in
                 completion(product, code)
             }
         } else {
@@ -413,7 +423,7 @@ extension ScannerViewController {
                 let startIndex = code.startIndex
                 let embeddedCode = String(code[code.index(startIndex, offsetBy: 2)..<code.index(startIndex, offsetBy: 15)])
                 let embeddedPrice = Int(String(code[code.index(startIndex, offsetBy: 16)..<code.index(startIndex, offsetBy: 21)])) ?? 0
-                self.productProvider.productByScannableCode(embeddedCode) { result, error in
+                self.productProvider.productByScannableCode(embeddedCode, self.shop.id) { result, error in
                     if let result = result {
                         let template = "2417000000000"
                         let newCode = EAN13.embedDataInEan(template, data: embeddedPrice)
@@ -423,7 +433,7 @@ extension ScannerViewController {
                     }
                 }
             } else {
-                self.productProvider.productByScannableCode(code) { result, error in
+                self.productProvider.productByScannableCode(code, self.shop.id) { result, error in
                     if let result = result {
                         completion(result.product, result.code)
                     } else {